requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
idna==3.10
jinja2==3.1.6
kiwisolver==1.4.8
lxml==5.4.0
markupsafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.2.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
requests==2.32.4
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing-extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.158
ultralytics-thop==2.0.14
urllib3==2.5.0
